The wallet is categories now: history, leaderboard, daily pool, store
One scroll held the balance, a four-tab board, two tiles, two pickers, the
gift, the shop and a disclaimer, so the thing you came for sat under three
things you did not. The landing page keeps what you have and the hourly gift
— the one thing in the wallet that asks for nothing — and a row per category
opens its own page, the shape Settings already has. The pickers live with the
store, above the prices, because what you own is the answer to why buy
another. The old /tokens screen is split between history and daily pool.
