+++ apps/mobile/src/i18n/messages/ar.ts
@@ -1235,6 +1235,10 @@ export const ar: Localized<EnMessages> = {
     streakFreezes: 'تجميدات السلسلة',
     itemsOwned: 'العناصر المملوكة',
     storeTitle: 'المتجر',
+    historyBody: 'ما كسبته وأنفقته يومًا بيوم',
+    leaderboardBody: 'من كسب الأكثر هذا الأسبوع وهذا الشهر وهذا العام',
+    poolBody: 'كيف تُحسب الحصة الليلية',
+    storeBody: 'تجميد السلسلة والأيام الفائتة والمظاهر',
     disclaimer:
       'الرموز نقاط داخل التطبيق. لا يمكن شراؤها أو تداولها أو سحبها أو استخدامها لفتح خطة مدفوعة — فقط تجميدات السلسلة والأيام الفائتة والعناصر التجميلية. لا سلسلة ولا عقد ولا سوق.',
   },
