+++ apps/mobile/src/i18n/messages/en.ts
@@ -1116,6 +1116,10 @@ export const en = {
     streakFreezes: 'Streak freezes',
     itemsOwned: 'Items owned',
     storeTitle: 'Store',
+    historyBody: 'What you earned and spent, day by day',
+    leaderboardBody: 'Who earned the most this week, month and year',
+    poolBody: 'How the nightly share is worked out',
+    storeBody: 'Streak freezes, missed days and cosmetics',
     disclaimer:
       'Tokens are in-app points. They cannot be bought, traded, withdrawn or used to unlock a paid plan — only streak freezes, missed days and cosmetics. There is no chain, no contract and no market.',
   },
