+++ apps/mobile/src/i18n/messages/fr.ts
@@ -1079,6 +1079,10 @@ export const fr: Localized<EnMessages> = {
     streakFreezes: 'Gels de série',
     itemsOwned: 'Objets possédés',
     storeTitle: 'Boutique',
+    historyBody: 'Ce que tu as gagné et dépensé, jour par jour',
+    leaderboardBody: 'Qui a gagné le plus cette semaine, ce mois et cette année',
+    poolBody: 'Comment la part de la nuit est calculée',
+    storeBody: 'Gels de série, jours manqués et cosmétiques',
     disclaimer:
       'Les jetons sont des points internes à l’app. Ils ne peuvent être achetés, échangés, retirés ni servir à débloquer une formule payante — seulement des gels de série, des jours manqués et des cosmétiques. Pas de chaîne, pas de contrat, pas de marché.',
   },
