+++ apps/mobile/src/i18n/messages/pt-BR.ts
@@ -1067,6 +1067,10 @@ export const ptBR: Localized<EnMessages> = {
     streakFreezes: 'Congelamentos de sequência',
     itemsOwned: 'Itens',
     storeTitle: 'Loja',
+    historyBody: 'O que você ganhou e gastou, dia a dia',
+    leaderboardBody: 'Quem ganhou mais nesta semana, neste mês e neste ano',
+    poolBody: 'Como a parte da noite é calculada',
+    storeBody: 'Congelamentos de sequência, dias perdidos e cosméticos',
     disclaimer:
       'As fichas são pontos dentro do app. Não podem ser compradas, trocadas, sacadas nem usadas para desbloquear um plano pago — apenas congelamentos de sequência, dias perdidos e cosméticos. Não há cadeia, contrato nem mercado.',
   },
