+++ apps/mobile/src/i18n/messages/ru.ts
@@ -1191,6 +1191,10 @@ export const ru: Localized<EnMessages> = {
     streakFreezes: 'Заморозки серии',
     itemsOwned: 'Предметы',
     storeTitle: 'Магазин',
+    historyBody: 'Что вы заработали и потратили, день за днём',
+    leaderboardBody: 'Кто заработал больше всех за неделю, месяц и год',
+    poolBody: 'Как считается ночная доля',
+    storeBody: 'Заморозки серии, пропущенные дни и косметика',
     disclaimer:
       'Жетоны — это внутренние баллы. Их нельзя купить, обменять, вывести или разблокировать ими платный тариф — только заморозки серии, пропущенные дни и косметика. Нет ни цепочки, ни контракта, ни рынка.',
   },
